+++ R/misc.R
@@ -58,12 +58,12 @@ tabs_as_spaces <- function(
   warn=getOption('fansi.warn', TRUE), ctl='all'
 ) {
   ## modifies / creates NEW VARS in fun env
-  if(!is.numeric(tab.stops) || !length(tab.stops) || any(tab.stops < 1))
-    stop("Argument `tab.stops` must be numeric and strictly positive")
-  VAL_IN_ENV(x=x, warn=warn, ctl=ctl, warn.mask=set_bits(5, 7, 9))
+  VAL_IN_ENV(
+    x=x, warn=warn, ctl=ctl, warn.mask=set_bits(5, 7, 9), tab.stops=tab.stops
+  )
   term.cap.int <- 1L
   .Call(
-    FANSI_tabs_as_spaces, x, as.integer(tab.stops), WARN.INT,
+    FANSI_tabs_as_spaces, x, tab.stops, WARN.INT,
     term.cap.int, CTL.INT
   )
 }
@@ -177,7 +177,7 @@ html_esc <- function(x, what=getOption("fansi.html.esc", "<>&'\"")) {
     stop("Argument `x` must be character, is ", typeof(x), ".")
   if(!is.character(what))
     stop("Argument `what` must be character, is ", typeof(what), ".")
-  .Call(FANSI_esc_html, enc2utf8(x), what)
+  .Call(FANSI_esc_html, enc_to_utf8(x), what)
 }
 
 #' Format Character Vector for Display as Code in HTML