+++ R/sgr.R
@@ -64,7 +64,7 @@ strip_ctl <- function(x, ctl='all', warn=getOption('fansi.warn', TRUE), strip) {
   ## modifies / creates NEW VARS in fun env
   VAL_IN_ENV(x=x, ctl=ctl, warn=warn, warn.mask=set_bits(5, 7, 9))
 
-  if(length(ctl)) .Call(FANSI_strip_csi, enc2utf8(x), CTL.INT, WARN.INT)
+  if(length(ctl)) .Call(FANSI_strip_csi, enc_to_utf8(x), CTL.INT, WARN.INT)
   else x
 }
 #' Strip Control Sequences
@@ -144,6 +144,7 @@ has_sgr <- function(x, warn=getOption('fansi.warn', TRUE))
 #'
 #' @export
 #' @inheritParams substr_ctl
+#' @inheritSection substr_ctl Output Stability
 #' @inherit has_ctl seealso
 #' @return character vector same length as `x`.
 #' @examples
@@ -195,6 +196,6 @@ close_state <- function(
 
 process <- function(x, ctl="all")
   .Call(
-    FANSI_process, enc2utf8(x), 1L, match(ctl, VALID.CTL)
+    FANSI_process, enc_to_utf8(x), 1L, match(ctl, VALID.CTL)
   )