+++ R/strsplit.R
@@ -36,6 +36,7 @@
 #' @inheritParams base::strsplit
 #' @inheritParams strwrap_ctl
 #' @inherit substr_ctl seealso
+#' @inheritSection substr_ctl Output Stability
 #' @note Non-ASCII strings are converted to and returned in UTF-8 encoding.
 #'   Width calculations will not work properly in R < 3.2.2.
 #' @inherit base::strsplit return
@@ -61,7 +62,7 @@ strsplit_ctl <- function(
     carry=carry, terminate=terminate
   )
   if(is.null(split)) split <- ""
-  split <- enc2utf8(as.character(split))
+  split <- enc_to_utf8(as.character(split))
   if(!length(split)) split <- ""
   if(anyNA(split)) stop("Argument `split` may not contain NAs.")
   if(any(Encoding(split) == "bytes"))